update store structure
