+++ src/ugd_frontend/src/components/ui/AuthButton.tsx
@@ -1,4 +1,5 @@
-import { useAppDispatch, useAppSelector } from "@/store/hooks";
+import { useAppDispatch } from "@/store/hooks/useAppDispatch";
+import { useAppSelector } from "@/store/hooks/useAppSelector";
 import React from "react";
 
 export default function AuthButton() {
