+++ src/ugd_frontend/src/components/ui/FilterButton.tsx
@@ -1,5 +1,6 @@
 import { setFilter } from "@/features/home/homeSlice";
-import { useAppDispatch, useAppSelector } from "@/store/hooks";
+import { useAppDispatch } from "@/store/hooks/useAppDispatch";
+import { useAppSelector } from "@/store/hooks/useAppSelector";
 import React from "react";
 
 export default function FilterButton() {
