+++ src/ugd_frontend/src/components/ui/SearchField.tsx
@@ -1,10 +1,9 @@
 import { setSearch, setView } from "@/features/home/homeSlice";
-import { useAppDispatch, useAppSelector } from "@/store/hooks";
+import { useAppDispatch } from "@/store/hooks/useAppDispatch";
+import { useAppSelector } from "@/store/hooks/useAppSelector";
 import React, {
 	ChangeEvent,
-	ChangeEventHandler,
 	KeyboardEvent,
-	useState,
 } from "react";
 import { IoIosSearch } from "react-icons/io";
 import { RxCross1 } from "react-icons/rx";
