+++ src/ugd_frontend/src/pages/Home/Categories/BookCarousel/index.tsx
@@ -4,8 +4,9 @@ import { Swiper, SwiperSlide } from "swiper/react";
 import { Navigation } from "swiper/modules";
 import "swiper/css";
 import "swiper/css/navigation";
-import { useAppDispatch, useAppSelector } from "@/store/hooks";
 import { setSelectedBook } from "@/features/home/homeSlice";
+import { useAppSelector } from "@/store/hooks/useAppSelector";
+import { useAppDispatch } from "@/store/hooks/useAppDispatch";
 // Sample data for the books
 const books = [
 	{
