+++ src/ugd_frontend/src/pages/Home/Categories/BookModal/index.tsx
@@ -1,6 +1,6 @@
 
 import { setSelectedBook } from "@/features/home/homeSlice";
-import { useAppDispatch } from "@/store/hooks";
+import { useAppDispatch } from "@/store/hooks/useAppDispatch";
 import React from "react";
