+++ src/ugd_frontend/src/store/hooks/useAppDispatch.ts
@@ -1,0 +1,4 @@
+import { useDispatch } from "react-redux";
+import { AppDispatch } from "../storeTypes";
+
+export const useAppDispatch: () => AppDispatch = useDispatch;
