+++ src/ugd_frontend/src/store/hooks/useAppSelector.ts
@@ -1,0 +1,5 @@
+import { useSelector } from "react-redux";
+import type { TypedUseSelectorHook } from "react-redux";
+import { RootState } from "../storeTypes";
+
+export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
