+++ src/ugd_frontend/src/store/hooks/useAppStore.ts
@@ -1,0 +1,4 @@
+import { AppStore } from "../storeTypes";
+import { useStore } from "react-redux";
+
+export const useAppStore: () => AppStore = useStore;
