+++ src/ugd_frontend/src/store/index.ts
@@ -1,16 +1,4 @@
 import { configureStore } from "@reduxjs/toolkit";
-import portalReducer from "@/features/portal/portalSlice";
-import homeReducer from "@/features/home/homeSlice";
+import rootReducer from "./rootReducer";
 
-export const makeStore = () => {
-	return configureStore({
-		reducer: {
-			portal: portalReducer,
-			home: homeReducer,
-		}
-	});
-};
-
-export type AppStore = ReturnType<typeof makeStore>;
-export type RootState = ReturnType<AppStore["getState"]>;
-export type AppDispatch = AppStore["dispatch"];
+export const makeStore = () => configureStore({ reducer: rootReducer })
