+++ src/ugd_frontend/src/store/rootReducer.ts
@@ -1,0 +1,10 @@
+import { combineReducers } from "@reduxjs/toolkit";
+import portalReducer from "@/features/portal/portalSlice";
+import homeReducer from "@/features/home/homeSlice";
+
+const rootReducer = combineReducers({
+	portal: portalReducer,
+	home: homeReducer,
+});
+
+export default rootReducer;
