+++ src/ugd_frontend/src/store/storeTypes.ts
@@ -1,0 +1,5 @@
+import { makeStore } from ".";
+
+export type AppStore = ReturnType<typeof makeStore>;
+export type RootState = ReturnType<AppStore["getState"]>;
+export type AppDispatch = AppStore["dispatch"];
